chanage SM4
