+++ sm3/ifile
@@ -1,0 +1,1 @@
+test
