+++ sm4/sm4.go
@@ -20,11 +20,7 @@ package sm4
 
 import (
 	"crypto/cipher"
-	"crypto/rand"
-	"crypto/x509"
-	"encoding/pem"
 	"errors"
-	"io/ioutil"
 	"strconv"
 )
 
@@ -241,66 +237,6 @@ func DecryptBlock(key SM4Key, dst, src []byte) {
 	cryptBlock(subkeys, make([]uint32, 4), make([]byte, 16), dst, src, true)
 }
 
-func ReadKeyFromMem(data []byte, pwd []byte) (SM4Key, error) {
-	block, _ := pem.Decode(data)
-	if block == nil {
-		return nil, errors.New("SM4: pem decode failed")
-	}
-	if x509.IsEncryptedPEMBlock(block) {
-		if block.Type != "SM4 ENCRYPTED KEY" {
-			return nil, errors.New("SM4: unknown type")
-		}
-		if pwd == nil {
-			return nil, errors.New("SM4: need passwd")
-		}
-		data, err := x509.DecryptPEMBlock(block, pwd)
-		if err != nil {
-			return nil, err
-		}
-		return data, nil
-	}
-	if block.Type != "SM4 KEY" {
-		return nil, errors.New("SM4: unknown type")
-	}
-	return block.Bytes, nil
-}
-
-func ReadKeyFromPem(FileName string, pwd []byte) (SM4Key, error) {
-	data, err := ioutil.ReadFile(FileName)
-	if err != nil {
-		return nil, err
-	}
-	return ReadKeyFromMem(data, pwd)
-}
-
-func WriteKeytoMem(key SM4Key, pwd []byte) ([]byte, error) {
-	if pwd != nil {
-		block, err := x509.EncryptPEMBlock(rand.Reader,
-			"SM4 ENCRYPTED KEY", key, pwd, x509.PEMCipherAES256)
-		if err != nil {
-			return nil, err
-		}
-		return pem.EncodeToMemory(block), nil
-	} else {
-		block := &pem.Block{
-			Type:  "SM4 KEY",
-			Bytes: key,
-		}
-		return pem.EncodeToMemory(block), nil
-	}
-}
-
-func WriteKeyToPem(FileName string, key SM4Key, pwd []byte) error {
-	pemBytes, err := WriteKeytoMem(key, pwd)
-	if err != nil {
-		return err
-	}
-	err = ioutil.WriteFile(FileName, pemBytes, 0666)
-	if err != nil {
-		return err
-	}
-	return nil
-}
 
 // NewCipher creates and returns a new cipher.Block.
 func NewCipher(key []byte) (cipher.Block, error) {
